+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Changelog
 - v5.3.0
     - Introduced shared library for extended telemetry to identify and prepare testing platform for native rust extensions.
+    - Fixed TIMESTAMP_LTZ datatype to honor session TIMEZONE parameter (ALTER SESSION SET TIMEZONE) instead of using local machine timezone.
 - v5.2.1
     - Bug fix: Fix the extremely rare case where intermittent network issues during uploads to Azure Blob Storage prevent metadata updates
 - v5.2.0